transfer cnotdihedral synthesis from quantum_info to synthesis
