+++ qiskit/synthesis/cnotdihedral/__init__.py
@@ -1,0 +1,16 @@
+# This code is part of Qiskit.
+#
+# (C) Copyright IBM 2017, 2018.
+#
+# This code is licensed under the Apache License, Version 2.0. You may
+# obtain a copy of this license in the LICENSE.txt file in the root directory
+# of this source tree or at http://www.apache.org/licenses/LICENSE-2.0.
+#
+# Any modifications or derivative works of this code must retain this
+# copyright notice, and modified files need to carry a notice indicating
+# that they have been altered from the originals.
+
+"""Module containing cnot-dihedral circuit synthesis."""
+
+from .cnotdihedral_decompose_2_qubits import decompose_cnotdihedral_2_qubits
+from .cnotdihedral_decompose_general import decompose_cnotdihedral_general
